0 warnings

+++ RuamEngine/RuamCore/Time/RuamTime.cpp
@@ -3,8 +3,8 @@
 namespace ruamTime
 {
 
-	float Time::m_currentFrame = 0.0f;
-	float Time::m_lastFrame = 0.0f;
+	double Time::m_currentFrame = 0.0f;
+	double Time::m_lastFrame = 0.0f;
 
 	void Time::Update()
 	{
